Added a Modal and a networked image
Changed the image loaded with the places in the places listfrom a static local image to a networked image. Added PlaceDetail component to manage a modal. When a place list item is pressed it is now selected and not deleted. The modal buttons will close the modal and delete the modal respectively. Pressing the back button when the modal is up will close the modal.